refatorizacion de la pagina comments con creacion de 4 nuevos componentes.
